+++ src/main/java/subway/persistence/jdbcDao/LineDao.java
@@ -1,4 +1,4 @@
-package subway.persistence.dao;
+package subway.persistence.jdbcDao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
